\name{djstock}
\alias{djstock}
\docType{data}
\title{Data on Dow Jones stock}
\description{Closing price of Dow Jones companies from 20 March 2019 to 9 June 2022.}
\usage{data(djstock)}
\format{
  Object of class \code{data.frame} with a total of 814 observations on 31 variables.
  The first variable represents the date of observation, while the subsequent variables represent the closing price of each Dow Jones company.
}
%\note{}
%\references{}